# Simplify inventory checks