--- a/crates/volta-core/src/inventory/mod.rs
+++ b/crates/volta-core/src/inventory/mod.rs
@@ -1,84 +1,87 @@
 //! Provides types for working with Volta's _inventory_, the local repository
 //! of available tool versions.
 
-mod node;
-mod package;
-mod yarn;
-
 use std::collections::BTreeSet;
 use std::path::Path;
 
 use crate::error::ErrorDetails;
 use crate::fs::read_dir_eager;
+use crate::layout::volta_home;
+use crate::tool::PackageConfig;
 use crate::version::parse_version;
 use failure::ResultExt;
-use lazycell::LazyCell;
-use regex::Regex;
+use log::debug;
 use semver::Version;
 use volta_fail::Fallible;
+use walkdir::WalkDir;
 
-/// Lazily loaded inventory.
-pub struct LazyInventory {
-    inventory: LazyCell<Inventory>,
+/// Checks if a given Node version image is available on the local machine
+pub fn node_available(version: &Version) -> Fallible<bool> {
+    volta_home().map(|home| {
+        home.node_image_root_dir()
+            .join(version.to_string())
+            .exists()
+    })
 }
 
-impl LazyInventory {
-    /// Constructs a new `LazyInventory`.
-    pub fn init() -> LazyInventory {
-        LazyInventory {
-            inventory: LazyCell::new(),
-        }
-    }
-
-    /// Forces the loading of the inventory and returns an immutable reference to it.
-    pub fn get(&self) -> Fallible<&Inventory> {
-        self.inventory.try_borrow_with(Inventory::current)
-    }
-
-    /// Forces the loading of the inventory and returns a mutable reference to it.
-    pub fn get_mut(&mut self) -> Fallible<&mut Inventory> {
-        self.inventory.try_borrow_mut_with(Inventory::current)
-    }
+/// Collects a set of all Node versions fetched on the local machine
+pub fn node_versions() -> Fallible<BTreeSet<Version>> {
+    volta_home().and_then(|home| read_versions(home.node_image_root_dir()))
 }
 
-/// The inventory of locally available tool versions.
-pub struct Inventory {
-    pub node: node::Collection,
-    pub yarn: yarn::Collection,
-    pub packages: package::Collection,
+/// Checks if a given Yarn version image is available on the local machine
+pub fn yarn_available(version: &Version) -> Fallible<bool> {
+    volta_home().map(|home| home.yarn_image_dir(&version.to_string()).exists())
 }
 
-impl Inventory {
-    /// Returns the current inventory.
-    fn current() -> Fallible<Inventory> {
-        Ok(Inventory {
-            node: node::Collection::load()?,
-            yarn: yarn::Collection::load()?,
-            packages: package::Collection::load()?,
+/// Collects a set of all Yarn versions fetched on the local machine
+pub fn yarn_versions() -> Fallible<BTreeSet<Version>> {
+    volta_home().and_then(|home| read_versions(home.yarn_image_root_dir()))
+}
+
+/// Checks if a given package version image is available on the local machine
+pub fn package_available(name: &str, version: &Version) -> Fallible<bool> {
+    volta_home().map(|home| home.package_image_dir(name, &version.to_string()).exists())
+}
+
+/// Collects a set of all Package Configs on the local machine
+pub fn package_configs() -> Fallible<BTreeSet<PackageConfig>> {
+    let package_dir = volta_home()?.default_package_dir();
+
+    WalkDir::new(&package_dir)
+        .max_depth(1)
+        .into_iter()
+        // Ignore any items which didn't resolve as `DirEntry` correctly.
+        // There is no point trying to do anything with those, and no error
+        // we can report to the user in any case. Log the failure in the
+        // debug output, though
+        .filter_map(|entry| match entry {
+            Ok(dir_entry) => {
+                // Ignore directory entries.
+                if dir_entry.file_type().is_file() {
+                    Some(dir_entry.into_path())
+                } else {
+                    None
+                }
+            }
+            Err(e) => {
+                debug!("{}", e);
+                None
+            }
         })
-    }
+        .map(|file_path| PackageConfig::from_file(&file_path))
+        .collect::<Fallible<BTreeSet<PackageConfig>>>()
 }
 
 /// Reads the contents of a directory and returns the set of all versions found
-/// in the directory's listing by matching filenames against the specified regex
-/// and parsing the `version` named capture as a semantic version.
-///
-/// The regex should contain the `version` named capture by using the Rust regex
-/// syntax `?P<version>`.
-fn versions_matching(dir: &Path, re: &Regex) -> Fallible<BTreeSet<Version>> {
+/// in the directory's listing by parsing the directory names as semantic versions
+fn read_versions(dir: &Path) -> Fallible<BTreeSet<Version>> {
     let contents = read_dir_eager(dir).with_context(|_| ErrorDetails::ReadDirError {
-        dir: dir.to_path_buf(),
+        dir: dir.to_owned(),
     })?;
 
-    let versions = contents
-        .filter(|(_, metadata)| metadata.is_file())
-        .filter_map(|(entry, _)| {
-            let path = entry.path();
-            let file_name = path.file_name()?.to_string_lossy();
-            let captures = re.captures(&file_name)?;
-            parse_version(&captures["version"]).ok()
-        })
-        .collect();
-
-    Ok(versions)
+    Ok(contents
+        .filter(|(_, metadata)| metadata.is_dir())
+        .filter_map(|(entry, _)| parse_version(entry.file_name().to_string_lossy()).ok())
+        .collect())
 }