… crates/volta-core/src/inventory.rs
… crates/volta-core/src/inventory/node.rs
… crates/volta-core/src/inventory/package.rs
… crates/volta-core/src/inventory/yarn.rs