Closes #210 

Info
-----
Our current implementation of `Inventory` has a few problems with how it collects data:
* For Node & Yarn, it looks at the Tarball cache to determine if a package is available, not the actual tool image. This means that if either the cached tarball or the image are removed, Volta winds up in an irrecoverable state.
* While the `Inventory` object is loaded lazily, once we load it, we load everything. So even if all we need to know is, "Do we have one specific version of Node available?", we have to catalogue all the Node, Yarn, and Package versions before determining the answer to the question. Most often on the hot path, we only need to know about a couple specific tools, and not the whole of the inventory.
* Because the Inventory was stateful, we needed to keep it in sync as we were doing fetches, which opened to the possibility of it getting out of sync at some point.

Changes
-----
* Removed the `Inventory` and `LazyInventory` structs, replacing them with functions that answer a specific question.
* Added `node_available`, `yarn_available`, and `package_available` functions that check for a single, specific version of Node, Yarn, or a Package, respectively.
* Added `node_versions`, `yarn_versions`, and `package_configs` functions, which return a Set of all available versions for each tool (similar to how the previous `Collection` objects were built).
* Updated the logic for Node and Yarn to look at the image directory, instead of the cached tarballs, to determine if a tool is installed.
* Renamed `fetch_internal` in all of the Tool implementations to `ensure_fetched`, which more accurately reflects what the method does.

Notes
-----
* The logic for determining the list of installed Packages was left unchanged, as it depends on the existence of `PackageConfig` files, instead of the image directories (since a package can have multiple binaries).
* Since we are no longer caching the entire inventory, there is the possibility of duplicate IO calls. However, only `volta list` is currently using the full list checks, and everywhere else is only doing single `$tool_available` calls. Currently there aren't any paths that need multiple calls, but perhaps we should try to instrument that in tests so we don't get regressions going forward.
* Also cleaned up some unneeded `.clone()` calls in tests that produced Clippy warnings.