@charlespierce 2020-03-26
Super non-scientific, anecdotal test: Running `node --version` 200 times on my Mac gives me an average runtime of `16.7 ms +- 1.3 ms`, while doing the same with this change gives me an average runtime of `14.5 ms +- 1.2 ms`, so this seems to give a significant (though small in an absolute sense) improvement to the hot path execution.