@chriskrycho 2020-03-31
For a slightly more significant test—

Before:

```sh
$ hyperfine "node --version" --warmup 10
Benchmark #1: node --version
  Time (mean ± σ):      13.2 ms ±   0.7 ms    [User: 6.4 ms, System: 4.5 ms]
  Range (min … max):    11.8 ms …  16.1 ms    172 runs
```

After:

```sh
hyperfine "node --version" --warmup 100
Benchmark #1: node --version
  Time (mean ± σ):      13.6 ms ±   0.8 ms    [User: 6.4 ms, System: 4.8 ms]
  Range (min … max):    11.7 ms …  15.4 ms    162 runs
```

Doesn't seem like it makes a meaningful difference at least with that.

---

Methodology: ran it with the original, then copied the original out of its place and copied the new release build into its place.