@chriskrycho 2020-03-31
Follow-up: my methodology was wrong; didn't have the relevant files in place. After doing it correctly (`./dev/unix/volta-install.sh --release`):

```sh
hyperfine "node --version" --warmup 100
Benchmark #1: node --version
  Time (mean ± σ):      12.1 ms ±   0.3 ms    [User: 5.6 ms, System: 4.0 ms]
  Range (min … max):    11.6 ms …  13.4 ms    201 runs
```

Confirms @charlespierce's results almost exactly!